+++ CHANGELOG.md
@@ -221,5 +221,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 ## [2.5.4]
 
+### Added
+- Added a 'Clear' button in Auto Tests and Paste Requirements mode
+
 ### Changed
 - Minor bug fixes